avm2: Translate ExternalInterface to AS
