+++ core/src/avm2/globals/flash/external/ExternalInterface.as
@@ -1,0 +1,11 @@
+package flash.external
+{
+   public final class ExternalInterface
+   {
+      public static native function get available(): Boolean;
+      
+      public static native function addCallback(functionName: String, closure: Function) : void;
+      
+      public static native function call(functionName: String, ... arguments) : *;
+   }
+}
