+++ core/src/avm2/globals/globals.as
@@ -163,6 +163,8 @@ include "flash/media/StageVideoAvailabilityReason.as"
 include "flash/media/VideoCodec.as"
 include "flash/media/VideoStatus.as"
 
+include "flash/external/ExternalInterface.as"
+
 include "flash/net.as"
 include "flash/net/FileFilter.as"
 include "flash/net/FileReference.as"
